fix(cuda_infer): correct GPTQ dequant layout, f16 scales, routing, and prefill
- Fix GPTQ dequant kernel: transpose qweight/scales indexing to match
  safetensors layout ([in_dim//8, out_dim] storage), unpack qzeros from
  packed int32 nibbles instead of misinterpreting as bf16
- Fix scales dtype: GPTQ stores scales as float16, not bfloat16.
  Add f16_to_f32 conversion and upload as float32 to GPU
- Fix MoE routing: cpu_topk now excludes already-selected experts,
  preventing all 8 slots from selecting the same expert
- Add prompt prefill loop: feed all prompt tokens through 40 layers
  to build correct KV cache and GatedDeltaNet state before decoding
- Add --token-ids parameter to bypass C BPE tokenizer (has merge bug)
- Replace argmax with temperature+top-K+repetition penalty sampling
- extract_weights.py: auto-generate tensor_index.bin alongside weights
- Add PR summary to docs/superpowers/PR/

Verified: layer 0 linear attention matches PyTorch reference exactly.

+++ cuda_infer/kernels.h
@@ -9,8 +9,8 @@ extern "C" {
 #endif
 
 void cuda_dequant_matvec_gptq(
-    const uint32_t *d_qweight, const uint16_t *d_scales,
-    const uint16_t *d_qzeros, const float *d_x,
+    const uint32_t *d_qweight, const float *d_scales,
+    const uint32_t *d_qzeros, const float *d_x,
     float *d_out, int out_dim, int in_dim, int group_size,
     cudaStream_t stream);
